+++ sites/jeromefitzgerald.com/src/components/AppBar/MenuMobile.tsx
@@ -15,7 +15,7 @@ import {
   // StyledOverlay,
 } from '@jeromefitz/design-system/components'
 import {
-  darkTheme,
+  // darkTheme,
   keyframes,
   styled,
 } from '@jeromefitz/design-system/stitches.config'
@@ -29,13 +29,9 @@ import {
 } from '@radix-ui/react-icons'
 import { useTheme } from 'next-themes'
 import NextLink from 'next/link'
-import { useRouter } from 'next/router'
 import * as React from 'react'
 import { useSwipeable } from 'react-swipeable'
-import { useEffectOnce } from 'react-use'
-import { useSound } from 'use-sound'
 
-import { navigation } from '~config/navigation'
 import { useUI } from '~context/UI'
 
 const slideIn = keyframes({
@@ -62,96 +58,16 @@ const StyledLink = styled('a', Flex, {
   },
 })
 
-const MenuMobile = () => {
-  /**
-   * @question can we lift this and not duplicate
-   */
-  // const kbar = useKBar()
-  const [swipeDownTransitionTime, swipeDownTransitionTimeSet] =
-    React.useState('250ms')
-  const router = useRouter()
-  const { theme, setTheme } = useTheme()
-  // const toasts = useToast()
-  const { audio, toggleAudio } = useUI()
-
-  const [playBleep] = useSound('/static/audio/bleep.mp3', {
-    soundEnabled: audio,
-    volume: 0.25,
-  })
-  const [playDisableSound] = useSound('/static/audio/disable-sound.mp3', {
-    soundEnabled: true,
-    volume: 0.25,
-  })
-  const [playEnableSound] = useSound('/static/audio/enable-sound.mp3', {
-    soundEnabled: true,
-    volume: 0.25,
-  })
-
-  // const handleToast = (props) => {
-  //   const { title } = props
-  //   if (toasts && toasts.current) {
-  //     toasts.current.message({
-  //       duration: 2000,
-  //       text: `Routing to: ${title}`,
-  //       type: 'default',
-  //     })
-  //   }
-  // }
-
-  const handleRouteInternal = (url) => {
-    playBleep()
-    void router.push(url)
-  }
-
-  const handleRouteExternal = (url) => {
-    playBleep()
-    void window.open(url)
-  }
-
-  const handleToggleTheme = React.useCallback(() => {
-    const newTheme = theme === 'dark' ? 'light' : 'dark'
-    document.documentElement.classList.toggle(darkTheme.className)
-    document.documentElement.classList.toggle('light-theme')
-    document.documentElement.style.setProperty('color-scheme', newTheme)
-    setTheme(newTheme)
-    playBleep()
-  }, [playBleep, setTheme, theme])
-
-  const handleToggleAudio = React.useCallback(() => {
-    audio ? playDisableSound() : playEnableSound()
-    toggleAudio()
-  }, [audio, playDisableSound, playEnableSound, toggleAudio])
-
+const MenuMobile = ({ handleSelect, navigationNonMutated }) => {
   /**
    * @custom to sheet
    */
-  const [navigationNonMutated, navigationNonMutatedSet] = React.useState(null)
-  useEffectOnce(() => {
-    navigationNonMutatedSet(navigation)
-  })
+  const { theme } = useTheme()
+  const { audio } = useUI()
   const [open, openSet] = React.useState(false)
-
-  const handleSelect = (event, item) => {
-    // console.dir(`> handleSelect`)
-    // console.dir(event)
-    // console.dir(item)
-    // void handleToast({ title: item?.titleExtended ?? item?.title })
-    // @todo turn into function return
-    if (item?.type === 'url.internal' && !!item.url) {
-      void handleRouteInternal(item.url)
-    }
-    if (item?.type === 'url.external' && !!item.url) {
-      void handleRouteExternal(item.url)
-    }
-    if (item?.type === 'audio') {
-      void handleToggleAudio()
-    }
-    if (item?.type === 'theme') {
-      void handleToggleTheme()
-    }
-    // event.preventDefault()
-    if (item.id === 'settings-audio' || item.id === 'settings-theme') {
-      event.preventDefault()
+  const handleSelectInternal = (event, item) => {
+    void handleSelect(event, item)
+    if (item?.type === 'audio' || item?.type === 'theme') {
       return
     }
     void openSet(false)
@@ -160,6 +76,8 @@ const MenuMobile = () => {
   /**
    * @hack please instead do this correctly 🤣️
    */
+  const [swipeDownTransitionTime, swipeDownTransitionTimeSet] =
+    React.useState('250ms')
   const handlers = useSwipeable({
     // onSwiping: (event) => {
     //   if (event.dir === 'Down') {
@@ -279,7 +197,9 @@ const MenuMobile = () => {
                               align="center"
                               justify="start"
                               gap="2"
-                              onClick={(event) => handleSelect(event, section)}
+                              onClick={(event) =>
+                                handleSelectInternal(event, section)
+                              }
                             >
                               {section.icon && section.icon}
                               <Box as="span">{section.title}</Box>
@@ -312,7 +232,9 @@ const MenuMobile = () => {
                                     align="center"
                                     justify="start"
                                     gap="2"
-                                    onClick={(event) => handleSelect(event, item)}
+                                    onClick={(event) =>
+                                      handleSelectInternal(event, item)
+                                    }
                                   >
                                     {/* {item.icon && item.icon} */}
                                     {item.icon && icon}
@@ -374,7 +296,9 @@ const MenuMobile = () => {
                                     align="center"
                                     justify="start"
                                     gap="2"
-                                    onClick={(event) => handleSelect(event, item)}
+                                    onClick={(event) =>
+                                      handleSelectInternal(event, item)
+                                    }
                                   >
                                     {/* {item.icon && item.icon} */}
                                     {item.icon && icon}
@@ -430,7 +354,9 @@ const MenuMobile = () => {
                                     align="center"
                                     justify="start"
                                     gap="2"
-                                    onClick={(event) => handleSelect(event, item)}
+                                    onClick={(event) =>
+                                      handleSelectInternal(event, item)
+                                    }
                                   >
                                     {item.icon && item.icon}
                                     <Box as="div">
